refactor: Better split of code
* Created h and cpp files for classes which was inside API headers for
now.
