+++ sources/files.cmake
@@ -1,3 +1,4 @@
+
 set(UNIT_TESTS 
   ActionTest.cpp 
   SoftwareTransportTest.cpp 
@@ -8,4 +9,26 @@ if(NOT AVR)
   list(APPEND UNIT_TESTS SimulatorTest.cpp)
 endif()
 
+set(LIB_H
+  ./include/api/IHardware.h
+  ./include/api/Meta.h
+  ./include/api/Action.h
+  ./include/api/Actor.h
+  ./include/api/Entity.h
+  ./include/api/ITransport.h
+  ./include/hardware/simulator/Simulator.h
+  ./include/actors/Player.h
+  ./include/actors/Weapon.h
+  ./include/transports/Software.h
+  ./include/transports/Dummy.h
+  ./include/weapons/Vampire.h
+)
 
+set(LIB_CPP
+  ./src/api/IHardware.cpp
+  ./src/api/Actor.cpp
+  ./src/api/ITransport.cpp
+  ./src/hardware/simulator/Simulator.cpp
+  ./src/weapons/Vampire.cpp
+  ./src/transports/Software.cpp
+)
