+++ sources/src/transports/Software.cpp
@@ -1,0 +1,8 @@
+#include <transports/Software.h>
+
+transport::Software::Software() {
+}
+
+void transport::Software::Queue(Action const &action) {
+  _player->Accept(action);
+}
