+++ sources/tests/SoftwareTransportTest.cpp
@@ -1,9 +1,9 @@
-#include <api/Action.h>
-#include <api/Entity.h>
-#include <api/ITransport.h>
+#include <actors/Player.h>
 #include <api/Meta.h>
-#include <unity.h>
+#include <transports/Software.h>
 #include <weapons/Vampire.h>
+
+#include <unity.h>
 
 void Expect_action_to_be_sent_to_right_target();
 
@@ -22,7 +22,7 @@ void Expect_action_to_be_sent_to_right_target() {
   // Objective is:
   // a weapon hits player using primary
   // a weapon hits player using secondary
-  Actor::Player targetPlayer;
+  actors::Player targetPlayer;
   targetPlayer.SetEntityValue(Entity::Tag::Health, 20);
   transport::Software transport;
   Weapons::Vampire weapon(transport);
